% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{calculate_gmm}
\alias{calculate_gmm}
\title{Fit a Guassian Mixture Model from a document-feature matrix}
\usage{
calculate_gmm(input_dfm, k = 5, seed = NULL)
}
\arguments{
\item{input_dfm}{a rectr_dfm object}

\item{k}{integer, number of topics}

\item{seed}{integer, to set the random seed for reproducibility. When leave as NULL, no random seed is set.}
}
\value{
a rectr_model object. rectr_model$theta is the topic-assignment matrix
}
\description{
This function fits a guassian mixture model (GMM) from a document-feature matrix.
}
