% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{create_corpus}
\alias{create_corpus}
\title{Create a multilingual corpus}
\usage{
create_corpus(text_content, lang)
}
\arguments{
\item{text_content}{a vector of text content}

\item{lang}{a vector of lang, using lower case, e.g. "de"}
}
\value{
a quanteda corpus
}
\description{
This function generates a multilingual corpus.
}
