% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{dfm_filtering}
\alias{dfm_filtering}
\title{Filter a document-feature matrix for systematic language differences}
\usage{
dfm_filtering(input_dfm, dimension = 100)
}
\arguments{
\item{input_dfm}{dfm generated by dfm_boe()}

\item{dimension}{integer, the first singular value extracted.}
}
\value{
an rectr_dfm object
}
\description{
This function filters a document-feature matrix using singular value decomposition.
}
