% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{filter_dfm}
\alias{filter_dfm}
\title{Filter a document-feature matrix for systematic language differences}
\usage{
filter_dfm(input_dfm, k, corpus, multiplication_factor = 2,
  dimension = 100)
}
\arguments{
\item{input_dfm}{dfm generated by dfm_boe()}

\item{k}{integer, number of topics}

\item{corpus}{a multilingual corpus generated by create_corpus()}

\item{multiplication_factor}{integer, select k * mulitiplication_factor columns from the SVD-ed dfm.}

\item{dimension}{integer, the first singular value to be extracted.}
}
\value{
an rectr_dfm object
}
\description{
This function filters a document-feature matrix using singular value decomposition.
}
