% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{get_ft}
\alias{get_ft}
\title{Download fastText word embeddings from the internet}
\usage{
get_ft(lang = "en", topwords = 2e+05)
}
\arguments{
\item{lang}{character, which language of fastText word embeddings to download, e.g. "en"}

\item{topwords}{integer, trim the word embeddings based on the most frequent words, default to top 200000 words}
}
\value{
Nothing, a file is saved.
}
\description{
This function downloads fastText word embeddings from the internet, trims the embeddings and saves as an R serialized object.
}
