% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{print.rectr_dfm}
\alias{print.rectr_dfm}
\title{Print a rectr_dfm object}
\usage{
\method{print}{rectr_dfm}(rectr_dfm)
}
\arguments{
\item{rectr_dfm}{a rectr_dfm object}
}
\value{
nothing
}
\description{
This function prints useful information from an rectr_dfm object.
}
