% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{print.rectr_model}
\alias{print.rectr_model}
\title{Print a rectr_model object}
\usage{
\method{print}{rectr_model}(rectr_model)
}
\arguments{
\item{rectr_model}{a rectr_model object}
}
\value{
nothing
}
\description{
Print useful information about a rectr_model.
}
