% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{read_ft}
\alias{read_ft}
\title{Reading word embeddings downloaded with get_ft()}
\usage{
read_ft(lang = c("en", "de", "fr"))
}
\arguments{
\item{lang}{a vector of languages, e.g. c("en", "de")}
}
\value{
a list of word embeddings
}
\description{
This function reads fastText word embeddings downloaded with get_ft(). The file(s) must be in the working directory.
}
