% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectr.R
\name{transform_dfm_boe}
\alias{transform_dfm_boe}
\title{Generate a document-feature matrix using word embeddings}
\usage{
transform_dfm_boe(
  corpus,
  emb = NULL,
  .progress = TRUE,
  mode = "bert",
  noise = FALSE,
  remove_stopwords = TRUE,
  bert_sentence_tokenization = TRUE
)
}
\arguments{
\item{corpus}{a multilingual corpus generated by create_corpus()}

\item{emb}{a list of word embeddings loaded from read_ft()}

\item{.progress}{boolean, displaying a progress bar or not}

\item{mode}{character, either 'bert' or 'fasttext'}

\item{noise}{boolean, printing noise so that you know the transmation is progressing}

\item{remove_stopwords, }{boolean, whether or not to remove stopwords}
}
\value{
a rectr_dfm object
}
\description{
This function generates document-feature matrix (dfm) from a multilingual corpus.
}
